requirements.txt
streamlit>=1.35
anthropic>=0.28
requests>=2.31
beautifulsoup4>=4.12
google-cloud-bigquery>=3.13
